Move createNotebookAndEntry function back to tags.e2e.spec.js
